Allow player to be dealt a Texas Hold'em hand
Added compareRanks and getHand function declarations in deck_of_cards.h
file. Added compareRanks and getHand function implimentations in
deck_of_cards.c file. Added Player struct, and dealToPlayer function in
main.c file.

The code before only shuffled a deck of cards, with no players in the
game. I started with one player, and dealt two cards to them. The cards
are then converted into a 'hand', which orders the cards by rank and
compares the suits. This hand value will be used to impliment the hand
ranges function in the future.