Added Push/Fold mechanic and basic hand-range functionality
Added pushOrFold function to main.c. Updated main() to allow the player
to play basic poker game. Added the open hand-range for the UTG position
in ranges.h file.

The user will be able to make decisions when opening the executable.
Based off of their given cards, the player must make a decision to
'Push' (raise or call) or 'Fold' (forfeit the hand). The correct poker
hands that are worthy of a Push are stored in the ranges.h file. If the
player makes a correct push, they are rewarded with a point, stored in
'player->score'. If they push incorrectly, they are deducted a point.
Folding does not affect the score. This is to temporarily represent the
effective value in Texas Hold'em.